REFACTOR: note operation according to authenticate user

+++ FunDoNotes/src/services/note.service.js
@@ -1,8 +1,8 @@
 import Note from '../models/note.model';
 
 //get all notes
-export const getAllNotes = async () => {
-    const data = await Note.find();
+export const getAllNotes = async (user_id) => {
+    const data = await Note.find({user_id: user_id});
     return data;
 };
 
@@ -19,7 +19,8 @@ export const updateNote = async (_id, body) => {
     console.log("INPUT - note.service -> updateNote ----->", _id, body);
     const data = await Note.findByIdAndUpdate(
         {
-            _id
+            _id: _id,
+            user_id: body.user_id
         },
             body,
         {
@@ -31,16 +32,16 @@ export const updateNote = async (_id, body) => {
 };
 
 //delete single note
-export const deleteNote = async (id) => {
+export const deleteNote = async (id, user_id) => {
     console.log("INPUT - note.service -> deleteNote ----->", id);
-    await Note.findByIdAndDelete(id);
+    await Note.findByIdAndDelete({_id: id, user_id: user_id});
     return '';
 };
 
 //get single note
-export const getNote = async (id) => {
-    console.log("INPUT - note.service -> getNote ----->", id);
-    const data = await Note.findById(id);
+export const getNote = async (id, user_id) => {
+    console.log("INPUT - note.service -> getNote ----->", "_id: "+id, "user_id: "+user_id);
+    const data = await Note.findOne({_id: id, user_id: user_id});
     console.log("OUTPUT - note.service -> getNote ----->", data);    
     return data;
 };
